+++ eNMS/forms/fields.py
@@ -96,6 +96,10 @@ class PythonField(StringField):
         return super().__call__(*args, **kwargs)
 
 
+class CodeField(StringField):
+    pass
+
+
 class PasswordSubstitutionField(PasswordField):
     def __call__(self, *args, **kwargs):
         kwargs["style"] = "background-color: #e8f0f7"
@@ -114,6 +118,7 @@ class NoValidationSelectMultipleField(SelectMultipleField):
 
 field_types = {
     BooleanField: "bool",
+    CodeField: "code",
     DateField: "date",
     DictField: "dict",
     DictSubstitutionField: "dict",