+++ eNMS/static/services/netmiko_validation_service.js
@@ -3,13 +3,12 @@ global
 field: false
 */
 
-const serviceType = "netmiko_validation_service";
-
 // eslint-disable-next-line
 function job(id) {
-  field("auto_find_prompt", serviceType, id)
+  const serviceClass = "netmiko_validation_service";
+  field("auto_find_prompt", serviceClass, id)
     .on("change", function() {
-      field("expect_string", serviceType, id).prop("disabled", this.checked);
+      field("expect_string", serviceClass, id).prop("disabled", this.checked);
     })
     .change();
 }